pyproject.toml
[project]
name = "eight25-audit"
version = "0.1.0"
description = "Lightweight AI-powered single-URL website audit tool"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "httpx>=0.27.2",
  "beautifulsoup4>=4.12.3",
  "lxml>=5.3.0",
  "pydantic>=2.9.2",
  "python-dotenv>=1.0.1",
  "openai>=1.55.3",
  "truststore>=0.10.0",
  "flask>=3.0.0",
]

[project.scripts]
eight25-audit = "eight25_audit.cli:main"
eight25-audit-web = "eight25_audit.web:main"

[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
